refactor: rename signArchives to signArchivesMacOs
