+++ app/ide-desktop/lib/client/tasks/signArchivesMacOs.js
